requirements.txt
aiopg
aiohttp
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-aiopg
